moves generate_paths_from_DataService_dict to serializer module
