+++ src/pydase/data_service/state_manager.py
@@ -6,10 +6,8 @@ from typing import TYPE_CHECKING, Any, Optional, cast
 
 import pydase.units as u
 from pydase.data_service.data_service_cache import DataServiceCache
-from pydase.utils.helpers import (
-    generate_paths_from_DataService_dict,
-    get_nested_value_from_DataService_by_path_and_key,
-)
+from pydase.utils.helpers import get_nested_value_from_DataService_by_path_and_key
+from pydase.utils.serializer import generate_paths_from_DataService_dict
 
 if TYPE_CHECKING:
     from pydase import DataService
